requirements.txt
alembic==1.4.2
amqp==5.2.0
blinker==1.4
celery==5.1.0
cryptography==40.0.2
dnspython==1.16.0
dataclasses
dominate==2.5.1
email-validator==1.1.1
flasgger==0.9.7.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.9.0
Flask-Cors==3.0.8
Flask-Datepicker==0.13
Flask-Migrate==2.5.3
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
gevent==22.10.2
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.13.0
pyjwt==2.4.0
PyMySQL==0.9.3
python-editor==1.0.4
pytz==2020.1
raven==6.10.0
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.18
uWSGI==2.0.23
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0